Isolate SDL3 display query in a subprocess to avoid Kivy segfault
The settings app runs inside Kivy, which uses SDL2 for its OpenGL
context. Initialising SDL3 in the same process (SDL_Init/SDL_Quit)
tears down that context and segfaults on rendering.

Move the SDL3 display query into sdl_detect.py, a small child process
that returns the monitor list as JSON. This keeps SDL3 out of Kivy's
process entirely while preserving correct Wayland resolution detection
(3840x2160, not the XWayland-scaled 4096x2304).
